Refactor welcome flow to use controllers
