+++ Sources/Features/First Launch/DestinationViewController.swift
@@ -1,40 +1,47 @@
 import Cocoa
 
-@objc protocol WelcomeViewDelegate: class {
-  func welcomeView(_ view: WelcomeView,
-                   didTapGetStarted button: NSButton)
+@objc protocol DestinationViewControllerDelegate: class {
+  func destinationViewController(_ controller: DestinationViewController,
+                                 didTapDirectoryButton button: NSButton)
+  func destinationViewController(_ controller: DestinationViewController,
+                                 didTapNextButton button: NSButton)
 }
 
-class WelcomeView: NSView {
-  weak var delegate: WelcomeViewDelegate?
+class DestinationViewController: ViewController {
+  weak var delegate: DestinationViewControllerDelegate?
 
   lazy var iconView = NSImageView()
   lazy var titleLabel = BoldLabel()
-  lazy var button = NSButton(title: "Get started", target: self, action: #selector(getStarted(_:)))
+  lazy var nextButton = NSButton(title: "Next", target: self, action: #selector(nextScreen(_:)))
   lazy var gridView = NSGridView()
+  lazy var directoryLabel = Label.init(labelWithString: "No destination")
+  lazy var directoryButton = NSButton(title: "Select directory",
+                                      target: self,
+                                      action: #selector(selectDirectory(_:)))
 
-  private var layoutConstraints = [NSLayoutConstraint]()
-
-  override init(frame frameRect: NSRect) {
-    super.init(frame: frameRect)
-    loadView()
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    configureViews()
   }
 
-  required init?(coder decoder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-
-  func reset() {
+  override func viewWillAppear() {
+    super.viewWillAppear()
     titleLabel.alphaValue = 0.0
     iconView.alphaValue = 0.0
-    button.alphaValue = 0.0
+    nextButton.alphaValue = 0.0
     gridView.alphaValue = 0.0
+    directoryButton.alphaValue = 0.0
+  }
+
+  override func viewDidAppear() {
+    super.viewDidAppear()
+    animateIn()
   }
 
   func animateIn() {
     CATransaction.begin()
     defer { CATransaction.commit() }
-    NSAnimationContext.current.duration = 1.5
+    NSAnimationContext.current.duration = 1.0
 
     do {
       let groupAnimation = CAAnimationGroup()
@@ -46,7 +53,6 @@ class WelcomeView: NSView {
       let position = CABasicAnimation(keyPath: "position.y")
       position.fromValue = 600
       groupAnimation.animations = [position, fadeIn]
-      titleLabel.layer?.anchorPoint = .init(x: 0.5, y: 0.5)
       titleLabel.layer?.add(groupAnimation, forKey: nil)
       titleLabel.alphaValue = 1.0
     }
@@ -61,9 +67,8 @@ class WelcomeView: NSView {
       let position = CABasicAnimation(keyPath: "position.y")
       position.fromValue = -600
       groupAnimation.animations = [position, fadeIn]
-      button.layer?.anchorPoint = .init(x: 0.5, y: 0.5)
-      button.layer?.add(groupAnimation, forKey: nil)
-      button.alphaValue = 1.0
+      directoryButton.layer?.add(groupAnimation, forKey: nil)
+      directoryButton.alphaValue = 1.0
     }
 
     do {
@@ -118,8 +123,22 @@ class WelcomeView: NSView {
       let position = CABasicAnimation(keyPath: "position.y")
       position.toValue = -600
       groupAnimation.animations = [position, fadeIn]
-      button.layer?.add(groupAnimation, forKey: nil)
-      button.alphaValue = 0.0
+      directoryButton.layer?.add(groupAnimation, forKey: nil)
+      directoryButton.alphaValue = 0.0
+    }
+
+    do {
+      let groupAnimation = CAAnimationGroup()
+      groupAnimation.isRemovedOnCompletion = false
+      groupAnimation.timingFunction = .init(name: CAMediaTimingFunctionName.easeInEaseOut)
+      let fadeIn = CABasicAnimation(keyPath: "opacity")
+      fadeIn.fromValue = 1
+      fadeIn.toValue = 0
+      let position = CABasicAnimation(keyPath: "position.y")
+      position.toValue = -600
+      groupAnimation.animations = [position, fadeIn]
+      nextButton.layer?.add(groupAnimation, forKey: nil)
+      nextButton.alphaValue = 0.0
     }
 
     do {
@@ -149,38 +168,37 @@ class WelcomeView: NSView {
 
   // MARK: - Private methods
 
-  private func loadView() {
+  private func configureViews() {
     iconView.imageScaling = .scaleAxesIndependently
     iconView.wantsLayer = true
-    let image = NSImage(named: "AppIcon")
+    let image = NSImage(named: "Backup")
+    image?.isTemplate = true
     iconView.image = image
-    iconView.translatesAutoresizingMaskIntoConstraints = false
+    iconView.contentTintColor = NSColor.init(named: "Blue")
 
-    titleLabel.font = NSFont.boldSystemFont(ofSize: 48)
-    titleLabel.stringValue = "Welcome to Syncalicious"
-    titleLabel.textColor = NSColor(named: "Pompadour")
+    titleLabel.font = NSFont.boldSystemFont(ofSize: 32)
+    titleLabel.stringValue = "Give Syncalicious a home"
     titleLabel.maximumNumberOfLines = 2
     titleLabel.lineBreakMode = .byWordWrapping
     titleLabel.alignment = .left
-    titleLabel.translatesAutoresizingMaskIntoConstraints = false
     titleLabel.wantsLayer = true
 
-    button.wantsLayer = true
-    button.isBordered = false
-    button.layer?.backgroundColor = NSColor.init(named: "Green")?.cgColor
-    button.layer?.cornerRadius = 18
-    button.layer?.borderColor = NSColor.red.cgColor
-    button.layer?.borderWidth = 0
-    button.contentTintColor = NSColor.white
-    button.translatesAutoresizingMaskIntoConstraints = false
+    nextButton.wantsLayer = true
+    nextButton.isBordered = false
+    nextButton.layer?.backgroundColor = NSColor.init(named: "Green")?.cgColor
+    nextButton.layer?.cornerRadius = 18
+    nextButton.layer?.borderColor = NSColor.red.cgColor
+    nextButton.layer?.borderWidth = 0
+    nextButton.contentTintColor = NSColor.white
 
     gridView.wantsLayer = true
     gridView.xPlacement = .fill
-    gridView.translatesAutoresizingMaskIntoConstraints = false
+    gridView.rowSpacing = 30
 
     let syncIcon = NSImageView()
-    let syncLabel = Label.init(labelWithString: "Sync application preferences across all your macs")
-    syncLabel.maximumNumberOfLines = 2
+    // swiftlint:disable line_length
+    let syncLabel = Label.init(labelWithString: "To sync and make backups of your preferences files, Syncalicious needs a directory to use as a home.")
+    syncLabel.maximumNumberOfLines = -1
     syncLabel.lineBreakMode = .byWordWrapping
     syncIcon.image = NSImage.init(named: "Synced")
     syncIcon.contentTintColor = NSColor.init(named: "Green")
@@ -188,50 +206,66 @@ class WelcomeView: NSView {
     syncRow.yPlacement = .center
 
     let backupIcon = NSImageView()
-    let backupLabel = Label.init(labelWithString: "Backup your preferences files in case something goes wrong")
-    backupLabel.maximumNumberOfLines = 2
+    // swiftlint:disable line_length
+    let backupLabel = Label.init(labelWithString: "To sync preferences files across Macs, pick a directory that is synced to the cloud,\ne.g. iCloud, Dropbox or Google Drive.")
+    backupLabel.maximumNumberOfLines = -1
     backupLabel.lineBreakMode = .byWordWrapping
-    backupIcon.contentTintColor = NSColor.init(named: "Blue")
-    backupIcon.image = NSImage.init(named: "Backup")
+    backupIcon.image = NSImage.init(named: "Cloud")
     let backupRow = gridView.addRow(with: [backupIcon, backupLabel])
     backupRow.yPlacement = .center
 
-    addSubview(gridView)
-    addSubview(titleLabel)
-    addSubview(iconView)
-    addSubview(button)
+    let fileLabel = gridView.addRow(with: [NSView(), directoryLabel])
+    fileLabel.yPlacement = .center
+    directoryLabel.lineBreakMode = .byTruncatingMiddle
+    directoryLabel.font = NSFont.boldSystemFont(ofSize: 14)
+    directoryButton.wantsLayer = true
+
+    view.addSubview(gridView)
+    view.addSubview(titleLabel)
+    view.addSubview(iconView)
+    view.addSubview(directoryButton)
+    view.addSubview(nextButton)
 
     NSLayoutConstraint.deactivate(layoutConstraints)
     layoutConstraints = [
-      iconView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: -120),
-      iconView.bottomAnchor.constraint(equalTo: bottomAnchor, constant: 72),
+      iconView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: -120),
+      iconView.bottomAnchor.constraint(equalTo: view.bottomAnchor, constant: 72),
       iconView.widthAnchor.constraint(equalToConstant: 512),
       iconView.heightAnchor.constraint(equalToConstant: 512),
 
-      titleLabel.widthAnchor.constraint(lessThanOrEqualToConstant: 340),
-      titleLabel.topAnchor.constraint(equalTo: topAnchor, constant: 64),
-      titleLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -64),
+      titleLabel.widthAnchor.constraint(lessThanOrEqualToConstant: 385),
+      titleLabel.topAnchor.constraint(equalTo: view.topAnchor, constant: 64),
+      titleLabel.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -64),
 
       gridView.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 32),
       gridView.leadingAnchor.constraint(equalTo: titleLabel.leadingAnchor),
       gridView.trailingAnchor.constraint(equalTo: titleLabel.trailingAnchor),
+
+      directoryButton.widthAnchor.constraint(equalToConstant: 120),
+      directoryButton.heightAnchor.constraint(equalToConstant: 48),
+      directoryButton.topAnchor.constraint(equalTo: gridView.bottomAnchor),
+      directoryButton.leadingAnchor.constraint(equalTo: gridView.leadingAnchor, constant: 48),
 
       syncIcon.heightAnchor.constraint(equalToConstant: 48),
       syncIcon.widthAnchor.constraint(equalToConstant: 48),
       backupIcon.heightAnchor.constraint(equalToConstant: 48),
       backupIcon.widthAnchor.constraint(equalToConstant: 48),
 
-      button.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -64),
-      button.centerXAnchor.constraint(equalTo: titleLabel.centerXAnchor),
-      button.heightAnchor.constraint(equalToConstant: 36),
-      button.widthAnchor.constraint(equalToConstant: 120)
+      nextButton.bottomAnchor.constraint(equalTo: view.bottomAnchor, constant: -48),
+      nextButton.centerXAnchor.constraint(equalTo: titleLabel.centerXAnchor),
+      nextButton.heightAnchor.constraint(equalToConstant: 36),
+      nextButton.widthAnchor.constraint(equalToConstant: 120)
     ]
-    NSLayoutConstraint.activate(layoutConstraints)
+    NSLayoutConstraint.constrain(layoutConstraints)
   }
 
   // MARK: - Actions
 
-  @objc func getStarted(_ sender: NSButton) {
-    delegate?.welcomeView(self, didTapGetStarted: sender)
+  @objc func selectDirectory(_ sender: NSButton) {
+    delegate?.destinationViewController(self, didTapDirectoryButton: sender)
+  }
+
+  @objc func nextScreen(_ sender: NSButton) {
+    delegate?.destinationViewController(self, didTapNextButton: sender)
   }
 }
